refactor(log): add more error logs
